requirements.txt
colorama==0.4.4
commonmark==0.9.1
pyenchant==3.2.0
Pygments==2.8.1
rich==10.0.1
typing-extensions==3.7.4.3
